Disentangle dependencies from bug-detectors to core
The bug-detector module should use the normal user-facing API to report
findings, register callbacks and the like to verify that those work
correctly and be able to be used as examples.
