+++ packages/bug-detectors/index.ts
@@ -1,5 +1,4 @@
 /*
-
  * Copyright 2023 Code Intelligence GmbH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,63 +14,4 @@
  * limitations under the License.
  */
 
-import * as findings from "./findings";
-import * as fs from "fs";
-import * as path from "path";
 export { getBugDetectorConfiguration } from "./configuration";
-
-// Export user-facing API for writing custom bug detectors.
-export {
-	reportFinding,
-	getFirstFinding,
-	clearFirstFinding,
-	Finding,
-} from "./findings";
-
-// Global API for bug detectors that can be used by instrumentation plugins.
-export interface BugDetectors {
-	reportFinding: typeof findings.reportFinding;
-}
-
-export const bugDetectors: BugDetectors = {
-	reportFinding: findings.reportFinding,
-};
-
-// Filters out disabled bug detectors and prepares all the others for dynamic import.
-export function getFilteredBugDetectorPaths(
-	bugDetectorsDirectory: string,
-	disableBugDetectors: string[],
-): string[] {
-	const disablePatterns = disableBugDetectors.map(
-		(pattern: string) => new RegExp(pattern),
-	);
-	return (
-		fs
-			.readdirSync(bugDetectorsDirectory)
-			// The compiled "internal" directory contains several files such as .js.map and .d.ts.
-			// We only need the .js files.
-			// Here we also filter out bug detectors that should be disabled.
-			.filter((bugDetectorPath) => {
-				if (!bugDetectorPath.endsWith(".js")) {
-					return false;
-				}
-
-				// Dynamic imports need .js files.
-				const bugDetectorName = path.basename(bugDetectorPath, ".js");
-
-				// Checks in the global options if the bug detector should be loaded.
-				const shouldDisable = disablePatterns.some((pattern) =>
-					pattern.test(bugDetectorName),
-				);
-
-				if (shouldDisable) {
-					console.log(
-						`Skip loading bug detector "${bugDetectorName}" because of user-provided pattern.`,
-					);
-				}
-				return !shouldDisable;
-			})
-			// Get absolute paths for each bug detector.
-			.map((file) => path.join(bugDetectorsDirectory, file))
-	);
-}
