+++ packages/bug-detectors/package.json
@@ -16,7 +16,8 @@
 	"main": "dist/index.js",
 	"types": "dist/index.d.js",
 	"dependencies": {
-		"@jazzer.js/fuzzer": "*"
+		"@jazzer.js/core": "*",
+		"@jazzer.js/hooking": "*"
 	},
 	"devDependencies": {},
 	"engines": {
