+++ packages/bug-detectors/tsconfig.json
@@ -6,7 +6,7 @@
 	},
 	"references": [
 		{
-			"path": "../fuzzer"
+			"path": "../core"
 		},
 		{
 			"path": "../hooking"
